requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# ailotime.py: 13
astral >= 1.6.1

# run.py: 9,10,11
discord.py >= 0.16.12

# ailotime.py: 10
python_dateutil >= 2.7.3

# ailotime.py: 12
pytz >= 2018.5
